Changes to Ease JSON Handling and Merging JSON
You could use these two files. one as a HTML file with a form which takes data and gives a JSON. The python file would take all the json files and merge them together. You could simply create a table in MySQL and insert the merged file directly to add to the table.